# Optimize PropertyDescriptor with flags